Move parameter url logic back to Parameters
